File Dump 2/3/20
File dump of code as of February 3, 2020. Problem figuring out when or how the guest report knows that a local item has been changed (i.e guest added) and therefore should add it. Consider adding guest to an array in guest report and by using push to add guest to string.